icommons_ext_tools/requirements/local.txt
# local environment requirements

#includes the base.txt requirements needed in all environments
-r base.txt

# below are requirements specific to the local environment
django-debug-toolbar==2.1
django-sslserver==0.21
mock==2.0.0

faker==3.0.0
